refactor(validation): rename test_structure module to test_layout
Rename the production validation module and its tests for clearer behavior-oriented naming while preserving compatibility imports.

Closes #2362

Implemented-By: Codex
