+++ pyproject.toml
@@ -122,7 +122,7 @@ hephaestus-repository-evidence = "hephaestus.github.repository_evidence:reposito
 hephaestus-system-info = "hephaestus.system.info:main"
 hephaestus-download-dataset = "hephaestus.datasets.downloader:main"
 hephaestus-check-python-version = "hephaestus.validation.python_version:main"
-hephaestus-check-test-structure = "hephaestus.validation.test_structure:main"
+hephaestus-check-test-structure = "hephaestus.validation.test_layout:main"
 hephaestus-check-coverage = "hephaestus.validation.coverage:main"
 hephaestus-check-complexity = "hephaestus.validation.complexity:main"
 hephaestus-filter-audit = "hephaestus.validation.audit:main"
